Begin reworking events loading.  Added tObjectFactory.

+++ source/event/event_list.cc
@@ -39,19 +39,19 @@ class cEvent;
 //  cEventList
 /////////////////
 cEventList::cEventList( cEventFactoryManager* factory_manager,  cEventTriggers *triggers ) :
-  m_factory_manager( factory_manager ),
-  m_triggers( triggers ),
-  m_head(NULL),
-  m_tail(NULL),
-  m_current(NULL),
-  m_num_events(0)
+m_factory_manager( factory_manager ),
+m_triggers( triggers ),
+m_head(NULL),
+m_tail(NULL),
+m_current(NULL),
+m_num_events(0)
 {
 }
 
 cEventList::~cEventList() {
   /*
-  XXX:  modified by kaben.  DeleteAll is equivalent.
-  */
+   XXX:  modified by kaben.  DeleteAll is equivalent.
+   */
   //while( m_head != NULL ){
   //  m_current = m_head;
   //  m_head = m_head->GetNext();
@@ -59,17 +59,17 @@ cEventList::~cEventList() {
   //}
   DeleteAll();
 
-  delete m_factory_manager;
-  delete m_triggers;
+delete m_factory_manager;
+delete m_triggers;
 }
 
 bool
 cEventList::AddEvent( cEventTriggers::eTriggerVariable trigger,
-			   double start, double interval, double stop,
-			   const cString & name, const cString & args ){
+                      double start, double interval, double stop,
+                      const cString & name, const cString & args ){
   assert( m_factory_manager != NULL );
   cEvent *event = m_factory_manager->ConstructEvent( name, args );
-
+  
   ///// Adding Event to the list /////
   if( event != NULL ){
     InsertEvent(event, trigger, start, interval, stop);
@@ -80,11 +80,11 @@ cEventList::AddEvent( cEventTriggers::eTriggerVariable trigger,
 
 
 void cEventList::InsertEvent(cEvent *event,
-			     cEventTriggers::eTriggerVariable trigger,
-			     double start, double interval, double stop){
+                             cEventTriggers::eTriggerVariable trigger,
+                             double start, double interval, double stop){
   assert( event != NULL );
   cEventListEntry *entry = new cEventListEntry(event, trigger,
-						start, interval, stop);
+                                               start, interval, stop);
   // If there are no events in the list yet.
   if( m_tail == NULL ){
     assert( m_head == NULL );
@@ -104,7 +104,7 @@ void cEventList::InsertEvent(cEvent *event,
 
 void cEventList::Delete(cEventListEntry *entry){
   assert( entry != NULL );
-
+  
   if( entry->GetPrev() != NULL ){
     entry->GetPrev()->SetNext( entry->GetNext() );
   }
@@ -119,22 +119,22 @@ void cEventList::Delete(cEventListEntry *entry){
     assert( entry == m_tail );
     m_tail = entry->GetPrev();
   }
-
+  
   delete entry;
 }
 
 
 /*
-public manipulator used while rebuilding event list.
-XXX:  added by kaben.
-*/
+ public manipulator used while rebuilding event list.
+ XXX:  added by kaben.
+ */
 void cEventList::DeleteAll(void){
   while( m_head != NULL ){
     m_current = m_head;
     m_head = m_head->GetNext();
     delete m_head;
   }
-
+  
   m_head = 0;
   m_tail = 0;
   m_current = 0;
@@ -144,63 +144,63 @@ void cEventList::DeleteAll(void){
 
 void cEventList::Process(){
   double t_val = 0; // trigger value
-
+  
   // Iterate through all entrys in event list
   cEventListEntry * entry = m_head;
   while( entry != NULL ){
-
+    
     cEventListEntry * next_entry = entry->GetNext();
-
+    
     // Check trigger condition
-
+    
     // IMMEDIATE Events always happen and are always deleted
     if( entry->GetTrigger() == cEventTriggers::IMMEDIATE  ){
       //cerr<<"IMMEDIATE EVENT "<<event->GetName()<<endl;
       entry->GetEvent()->Process();
       Delete(entry);
     }else{
-
+      
       // Get the value of the appropriate trigger variable
       t_val = m_triggers->GetTriggerValue(entry->GetTrigger());
-
+      
       if( t_val != DBL_MAX &&
-	  ( t_val >= entry->GetStart() ||
-	    entry->GetStart() == cEventTriggers::TRIGGER_BEGIN ) &&
-	  ( t_val <= entry->GetStop() ||
-	    entry->GetStop() == cEventTriggers::TRIGGER_END ) ){
-	
-	entry->GetEvent()->Process();
-	
-	if( entry == NULL ){ // It is possible for an event to kill itself
-	}else{	
-	  // Handle the interval thing
-	  if( entry->GetInterval() == cEventTriggers::TRIGGER_ALL ){
-	    // Do Nothing
-	  }else if( entry->GetInterval() == cEventTriggers::TRIGGER_ONCE ){
-	    // If it is a onetime thing, remove it...
-	    Delete(entry);
-	  }else{
-	    // There is an interal.. so add it
-	    entry->NextInterval();
-	  }
-
-	  // If the event can never happen now... excize it
-	  if( entry != NULL  &&  entry->GetStop() != cEventTriggers::TRIGGER_END ){
-	    if( entry->GetStart() > entry->GetStop() &&
-		entry->GetInterval() > 0 ){
-	      Delete(entry);
-	    }else if( entry->GetStart() < entry->GetStop() &&
-		      entry->GetInterval() < 0 ){
-	      Delete(entry);
-	    }
-	  }
-	
-	}
-	
+          ( t_val >= entry->GetStart() ||
+            entry->GetStart() == cEventTriggers::TRIGGER_BEGIN ) &&
+          ( t_val <= entry->GetStop() ||
+            entry->GetStop() == cEventTriggers::TRIGGER_END ) ){
+        
+        entry->GetEvent()->Process();
+        
+        if( entry == NULL ){ // It is possible for an event to kill itself
+        }else{	
+          // Handle the interval thing
+          if( entry->GetInterval() == cEventTriggers::TRIGGER_ALL ){
+            // Do Nothing
+          }else if( entry->GetInterval() == cEventTriggers::TRIGGER_ONCE ){
+            // If it is a onetime thing, remove it...
+            Delete(entry);
+          }else{
+            // There is an interal.. so add it
+            entry->NextInterval();
+          }
+          
+          // If the event can never happen now... excize it
+          if( entry != NULL  &&  entry->GetStop() != cEventTriggers::TRIGGER_END ){
+            if( entry->GetStart() > entry->GetStop() &&
+                entry->GetInterval() > 0 ){
+              Delete(entry);
+            }else if( entry->GetStart() < entry->GetStop() &&
+                      entry->GetInterval() < 0 ){
+              Delete(entry);
+            }
+          }
+          
+        }
+        
       } // End Non-IMMEDITAE events
-
+      
     }  // end condition to do event
-
+    
     entry = next_entry;
   }
 }
@@ -221,29 +221,29 @@ void cEventList::Sync(){
 void cEventList::SyncEvent(cEventListEntry *entry){
   // Ignore events that are immdeiate
   if( entry->GetTrigger() == cEventTriggers::IMMEDIATE ){ return; }
-
+  
   double t_val = m_triggers->GetTriggerValue( entry->GetTrigger() );
-
+  
   // If t_val has past the end, remove (even if it is TRIGGER_ALL)
   if( t_val > entry->GetStop() ){
     Delete(entry);
     return;
   }
-
+  
   // If it is a trigger once and has passed, remove
   if( t_val > entry->GetStart() && entry->GetInterval() == cEventTriggers::TRIGGER_ONCE ){
     Delete(entry);
     return;
   }
-
+  
   // If for some reason t_val has been reset or soemthing, rewind
   if( t_val + entry->GetInterval() <= entry->GetStart() ){
     entry->Reset();
   }
-
+  
   // Can't fast forward events that are Triger All
   if( entry->GetInterval() == cEventTriggers::TRIGGER_ALL ){ return; }
-
+  
   // Keep adding interval to start until we are caught up
   while( t_val > entry->GetStart() ){
     entry->NextInterval();
@@ -262,11 +262,11 @@ void cEventList::PrintEventList( ostream & os ){
 }
 
 /*
-XXX:  modified by kaben to stream-dump in parseable format.
-*/
+ XXX:  modified by kaben to stream-dump in parseable format.
+ */
 void cEventList::PrintEvent(cEventListEntry * entry, ostream & os){
-    assert( entry != NULL );
-    switch ( entry->GetTrigger() ){
+  assert( entry != NULL );
+  switch ( entry->GetTrigger() ){
     case cEventTriggers::UPDATE:
       //os<<"UPDATE ";
       os<<"update ";
@@ -282,52 +282,52 @@ void cEventList::PrintEvent(cEventListEntry * entry, ostream & os){
     default:
       //os<<"UNDEFINED ";
       os<<"undefined ";
+  }
+  //os<<"[";
+  if (entry->GetTrigger() != cEventTriggers::IMMEDIATE ){
+    if( entry->GetStart() == cEventTriggers::TRIGGER_BEGIN ){
+      os<<"begin";
+    }else{
+      os<<entry->GetStart();
     }
-    //os<<"[";
-    if (entry->GetTrigger() != cEventTriggers::IMMEDIATE ){
-      if( entry->GetStart() == cEventTriggers::TRIGGER_BEGIN ){
-        os<<"begin";
-      }else{
-        os<<entry->GetStart();
-      }
-      os<<":";
-      if( entry->GetInterval() == cEventTriggers::TRIGGER_ONCE ){
-        os<<"once";
-      }else if( entry->GetInterval() == cEventTriggers::TRIGGER_ALL ){
-        os<<"all";
-      }else{
-        os<<entry->GetInterval();
-      }
-      os<<":";
-      if( entry->GetStop() == cEventTriggers::TRIGGER_END ){
-        os<<"end";
-      }else{
-        os<<entry->GetStop();
-      }
-      os<<" ";
+    os<<":";
+    if( entry->GetInterval() == cEventTriggers::TRIGGER_ONCE ){
+      os<<"once";
+    }else if( entry->GetInterval() == cEventTriggers::TRIGGER_ALL ){
+      os<<"all";
+    }else{
+      os<<entry->GetInterval();
     }
-    //os<<"] "<<entry->GetName()<<" "<<entry->GetArgs()<<endl;
-    os<<entry->GetName()<<" "<<entry->GetArgs()<<endl;
+    os<<":";
+    if( entry->GetStop() == cEventTriggers::TRIGGER_END ){
+      os<<"end";
+    }else{
+      os<<entry->GetStop();
+    }
+    os<<" ";
+  }
+  //os<<"] "<<entry->GetName()<<" "<<entry->GetArgs()<<endl;
+  os<<entry->GetName()<<" "<<entry->GetArgs()<<endl;
 }
 
 
 //// Parsing Event List File Format ////
 bool cEventList::AddEventFileFormat(const cString & in_line){
   cString cur_line = in_line;
-
+  
   // Timing
   cEventTriggers::eTriggerVariable trigger = cEventTriggers::UPDATE;
   double start = cEventTriggers::TRIGGER_BEGIN;
   double interval = cEventTriggers::TRIGGER_ONCE;
   double stop = cEventTriggers::TRIGGER_END;
-
+  
   cString name;
   cString arg_list;
-
+  
   cString tmp;
-
+  
   cString cur_word = cur_line.PopWord();
-
+  
   // Get the trigger variable if there
   if( cur_word == "i"  ||
       cur_word == "immediate" ){
@@ -335,35 +335,35 @@ bool cEventList::AddEventFileFormat(const cString & in_line){
     name = cur_line.PopWord();
     return AddEvent(name, cur_line); // If event is IMMEDIATE shortcut
   }else if( cur_word == "u"  ||
-	    cur_word == "update" ){
+            cur_word == "update" ){
     trigger = cEventTriggers::UPDATE;
     cur_word = cur_line.PopWord();
   }else if( cur_word == "g"  ||
-	    cur_word == "generation" ){
+            cur_word == "generation" ){
     trigger = cEventTriggers::GENERATION;
     cur_word = cur_line.PopWord();
   }else{
     // If Trigger is skipped so assume IMMEDIATE
     trigger = cEventTriggers::IMMEDIATE;
   }
-
+  
   // Do we now have timing specified?
   // Parse the Timing
   cString timing_str = cur_word;
-
+  
   // Get the start:interval:stop
   tmp = timing_str.Pop(':');
-
+  
   // If first value is valid, we are getting a timing.
   if( tmp.IsNumber() || tmp == "begin" ){
-
+    
     // First number is start
     if( tmp == "begin" ){
       start = cEventTriggers::TRIGGER_BEGIN;
     }else{
       start = tmp.AsDouble();
     }
-
+    
     // If no other words... is "start" syntax
     if( timing_str.GetSize() == 0 ){
       interval = cEventTriggers::TRIGGER_ONCE;
@@ -372,37 +372,37 @@ bool cEventList::AddEventFileFormat(const cString & in_line){
       // Second word is interval
       tmp = timing_str.Pop(':');
       if( tmp == "all" ){
-	  interval = cEventTriggers::TRIGGER_ALL;
+        interval = cEventTriggers::TRIGGER_ALL;
       }else if( tmp == "once" ){
-	interval = cEventTriggers::TRIGGER_ONCE;
+        interval = cEventTriggers::TRIGGER_ONCE;
       }else{
-	interval = tmp.AsDouble();
+        interval = tmp.AsDouble();
       }
       // If no other words... is "start:interval" syntax
       if( timing_str.GetSize() == 0 ){
-	stop     = cEventTriggers::TRIGGER_END;
+        stop     = cEventTriggers::TRIGGER_END;
       }else{
-	// We have "start:interval:stop" syntax
-	tmp = timing_str;
-	if( tmp == "end" ){
-	  stop = cEventTriggers::TRIGGER_END;
-	}else{
-	  stop = tmp.AsDouble();
-	}
+        // We have "start:interval:stop" syntax
+        tmp = timing_str;
+        if( tmp == "end" ){
+          stop = cEventTriggers::TRIGGER_END;
+        }else{
+          stop = tmp.AsDouble();
+        }
       }
     }
     cur_word = cur_line.PopWord(); // timing provided, so get next word
-
+    
   }else{ // We don't have timing, so assume IMMEDIATE
     trigger = cEventTriggers::IMMEDIATE;
     start = cEventTriggers::TRIGGER_BEGIN;
     interval = cEventTriggers::TRIGGER_ONCE;
     stop = cEventTriggers::TRIGGER_END;
   }
-
+  
   // Get the rest of the info
   name = cur_word;
   arg_list = cur_line;
-
+  
   return AddEvent( trigger, start, interval, stop, name, arg_list );
 }
